move interaction protocol to dedicated file
